CORGI refactoring + new features.
* Renamed "Components" to be "Systems"
  (Will rename "ComponentData" to "Component" soon!)
* You can (and should!) now explicitly declare any
  systems that your own system is dependent on.  There
  are several forms of dependency.  For now, it just
  lets new entities be automatically registered for
  any dependent systems when they are registered with
  the dependee.
  (In the future: enforcing execution order based on
  dependencies, and intelligent multithreading dispatch!)
* Misc changes to telegraph app for testing.

+++ telegram/src/states/base_state.h
@@ -6,8 +6,8 @@ class StateManager;
 class BaseState {
 public:
 
-  virtual void Update(float delta_time) {}
-  virtual void Render(float delta_time) {}
+  virtual void Update(double delta_time) {}
+  virtual void Render(double delta_time) {}
 
   virtual void Init() {}
   virtual void Cleanup() {}
